requirements.txt
aniso8601==0.85
crypto==1.4.1
cssmin==0.2.0
python-dateutil==2.2
Flask==0.10.1
Flask_Assets==0.10
Flask-compressor==0.2.0
flask_oauthlib==0.7.0
Flask-Restless==0.15.0
flask-restful==0.3.0
flask_socketio==0.6.0
Flask-SQLAlchemy==2.0
flask_wtf==0.10.2
google-api-python-client==1.3.1
gunicorn==19.1.1
httplib2==0.9
influxdb==0.4.1
itsdangerous==0.24
Jinja2==2.8
jsmin==2.1.3
PYjwt==1.4.1
MarkupSafe==0.23
mimerender==0.5.4
naked==0.1.31
oauth2client==1.4.1
pyasn1==0.1.9
pyasn1_modules==0.0.5
pycrypto==2.6.1
PyJWT==1.4.1
python_dateutil==2.2
python_mimeparse==0.1.4
pytz==2014.9
pyYAML==3.11
requests==2.4.3
rsa==3.4.2
shellescape==3.4.1
simplejson==3.6.5
six==1.8.0
sqlalchemy==0.9.8
uritemplate==0.6
webassets==0.10.1
Werkzeug==0.10.1
wsgiref==0.1.2
pyyaml==3.11






